+++ Frontend/src/pages/modes/exam-page/ExamPage.jsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import EntriesSection from '../../../components/entries-section/EntriesSection'
-import AuxSectionOne from '../../../components/aux-section-one/AuxSectionOne'
 import AuxSectionTwo from '../../../components/aux-section-two/AuxSectionTwo'
 import "./ExamPage.css"
+import HelpSection from '../../../components/aux-section-one/HelpSection'
 
 const ExamPage = () => {
   return (
     <div className='modes_page_container exam-color'>
       <p className='head_task'>Modo Examen</p>
       <EntriesSection />
-      <AuxSectionOne />
+      <HelpSection />
       <AuxSectionTwo />
     </div>
   )
