+++ Frontend/src/pages/modes/practice-page/PracticePage.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import AuxSectionTwo from '../../../components/aux-section-two/AuxSectionTwo'
-import AuxSectionOne from '../../../components/aux-section-one/AuxSectionOne'
 import EntriesSection from '../../../components/entries-section/EntriesSection'
 import "./PracticePage.css"
 import useScreenSize from '../../../hooks/useScreenSize'
 import EntryMobile from '../../../components/entries-section/entry-mobile/EntryMobile'
+import HelpSection from '../../../components/aux-section-one/HelpSection'
 
 const PracticePage = () => {
 
@@ -18,7 +18,7 @@ const PracticePage = () => {
     <div className='modes_page_container practice_color'>
       <h1 className='head_task'>Modo Tarea</h1>
       <EntriesSection />
-      <AuxSectionOne />
+      <HelpSection />
       <AuxSectionTwo />
     </div>
   )
